+++ src/index.js
@@ -3,6 +3,7 @@ import {
   onReady,
   onceReady,
   renderSidebarTab,
+  renderItemSheet,
   renderJournalSheet,
 } from "./hooks.js";
 import extendSceneNavigationContext from "./hooks/getSceneNavigationContext/extendSceneNavigationContext.js";
@@ -14,5 +15,6 @@ Hooks.once("ready", onceReady);
 Hooks.on("ready", onReady);
 Hooks.on("renderSidebarTab", renderSidebarTab);
 Hooks.on("renderJournalSheet", renderJournalSheet);
+Hooks.on("renderItemSheet", renderItemSheet);
 Hooks.on("getSceneNavigationContext", extendSceneNavigationContext);
 Hooks.on("getSceneDirectoryEntryContext", extendSceneNavigationContext);