+++ src/muncher/table.js
@@ -8,13 +8,13 @@ function diceRollMatcher(match, p1, p2, p3, p4, p5) {
     let dmgString = `${p4} damage`;
     dmgString = dmgString[0].toUpperCase() + dmgString.substring(1);
     const diceString = utils.parseDiceString(p2, null, `[${p4.toLowerCase()}]`).diceString;
-    return `${p1 ? p1: ""}[[/r ${diceString} # ${dmgString}]]${p3} damage`;
+    return `${p1 ? p1 : ""}[[/r ${diceString} # ${dmgString}]]${p3} damage`;
   } else if (p5 && p1 && p5.toLowerCase() === "points" && p1.toLowerCase() === "regains") {
     const diceString = utils.parseDiceString(p2, null, "[healing]").diceString;
-    return `${p1 ? p1: ""}[[/r ${diceString} # Healing]]${p3} hit points`;
+    return `${p1 ? p1 : ""}[[/r ${diceString} # Healing]]${p3} hit points`;
   } else {
     const diceString = utils.parseDiceString(p2).diceString;
-    const result = `${p1 ? p1: ""}[[/r ${diceString}]]${p3 ? p3 : ""} ${p4 ? p4 : ""} ${p5 ? p5 : ""}`.trim();
+    const result = `${p1 ? p1 : ""}[[/r ${diceString}]]${p3 ? p3 : ""} ${p4 ? p4 : ""} ${p5 ? p5 : ""}`.trim();
     return result;
   }
 }
@@ -36,7 +36,7 @@ function findDiceColumns(table) {
   if (table.tHead) {
     const headings = getHeadings(table);
     headings.forEach((h) => {
-      const diceRegex = new RegExp(/(\d*d\d+(\s*[+-]?\s*\d*)?)/, "g");
+      const diceRegex = new RegExp(/(\d*[d|D]\d+(\s*[+-]?\s*\d*)?)/, "g");
       const match = h.replace(/[­––−-]/gu, "-").replace(/-+/g, "-").match(diceRegex);
       if (match) {
         result.push(h);
@@ -46,10 +46,8 @@ function findDiceColumns(table) {
   return result;
 }
 
-function guessTableName(parentName, document, tableNum) {
-  // const element = document.querySelector(`table[data-content-chunk-id='${contentChunkId}']`);
-  const element = document.querySelectorAll('table');
-  // const element = $(`table:nth-child(${tableNum})`);
+function guessTableName(parentName, htmlDocument, tableNum) {
+  const element = htmlDocument.querySelectorAll('table');
   let track = element[tableNum];
   let sibling = track.previousElementSibling;
 
@@ -62,7 +60,6 @@ function guessTableName(parentName, document, tableNum) {
   }
 
   if (sibling) {
-    console.log(sibling.textContent);
     return sibling.textContent;
   } else {
     logger.warn(`No table name identified for ${parentName}`);
@@ -71,63 +68,22 @@ function guessTableName(parentName, document, tableNum) {
 }
 
 
-function sleep(ms) {
-  return new Promise((resolve) => {
-    setTimeout(resolve, ms);
-  });
+function tableReplacer(htmlDocument, tableNum, compendiumTables) {
+  // future enhancement - replace liks to DDB spells, monsters, items etc to munched compendium
+  const element = htmlDocument.querySelectorAll('table');
+  const tablePoint = element[tableNum];
+
+  if (tablePoint) {
+    const rollCompendiumLabel = game.settings.get("ddb-importer", "entity-table-compendium");
+    compendiumTables.slice().reverse().forEach((table) => {
+      logger.debug(`Updating table reference for: ${table.name}`);
+      tablePoint.insertAdjacentHTML("afterend", `<div id="table-link">@Compendium[${rollCompendiumLabel}.${table.name}]{Open RollTable ${table.name}}</div>`);
+    });
+   }
+
+  return htmlDocument;
 }
 
-function tableReplacer(text, journals) {
-  text = replacer.moduleReplaceLinks(text, journals, config);
-  text = replacer.foundryCompendiumReplace(text, config);
-  const document = utils.htmlToDoc(text);
-  text = document.body.textContent;
-}
-
-function textTableReplacer(text, tables) {
-  const dom = utils.htmlToDoc(text);
-  // document.querySelector('.titanic:nth-child(2)')
-
-  tables.forEach((table) => {
-    const tablePoint = dom.body.querySelector(`table[data-content-chunk-id="${table.flags.ddb.contentChunkId}"]`);
-    if (tablePoint) {
-      console.log(`Updating table reference for: ${table.name}`);
-      tablePoint.insertAdjacentHTML("afterend", `<div id="table-link">@RollTable[${table.name}]{Open RollTable}</div>`);
-    }
-  });
-  text = dom.body.innerHTML;
-}
-
-async function fixUpTables(tables, journals) {
-  console.log("Updating table references for modules...");
-  console.log(`There are ${tables.length} tables`);
-  console.log(`There are ${journals.length} journals`);
-
-  await sleep(1000);
-
-  for (let tableIndex = 0, tablesLength = tables.length; tableIndex < tablesLength; tableIndex++) {
-    const table = tables[tableIndex];
-    console.log(`Updating table: ${table.name}...`);
-    for (let resultsIndex = 0, resultsLength = table.results.length; resultsIndex < resultsLength; resultsIndex++) {
-      tableReplacer(table.results[resultsIndex].text, journals);
-    }
-    if (tableIndex % 10 == 0) {
-      await sleep(500);
-    }
-  }
-  await sleep(1000);
-
-  console.log("Starting Journal Table Updates");
-
-  for (let journalIndex = 0, journalsLength = journals.length; journalIndex < journalsLength; journalIndex++) {
-    journalTableReplacer(journals[journalIndex], tables);
-    if (journalIndex % 5 == 0) {
-      await sleep(500);
-    }
-  }
-
-  return [tables, journals];
-}
 
 function diceInt(text) {
   if (text === "0") return 10;
@@ -150,8 +106,6 @@ function getDiceTableRange(value) {
   const valueRegex = new RegExp(/^(\d+)\-(\d+)|^(\d+)(\+?)$/);
   const valueMatch = text.match(valueRegex);
 
-  console.warn(valueMatch);
-
   if (valueMatch) {
     if (valueMatch[1] !== undefined && valueMatch[2] !== undefined) {
       const low = diceInt(valueMatch[1]);
@@ -171,8 +125,7 @@ function getDiceTableRange(value) {
     }
   }
 
-  console.log(`Unable to table range match ${value}`);
-  console.log(`Text value: ${text}`);
+  // logger.debug(`Dice range: Unable to table range match '${value}' text was '${text}'`);
   return [];
 }
 
@@ -183,7 +136,7 @@ function buildTable(parsedTable, keys, diceKeys, tableName, parentName) {
   diceKeys.forEach((diceKey) => {
     const nameExtension = diceKeys > 1 ? ` [${diceKeys}]` : "";
     const realName = ((tableName && tableName !== "") ? tableName : "Unnamed Table") + nameExtension;
-    console.log(`Generating table ${realName}`);
+    logger.debug(`Generating table ${realName}`);
 
     const diceRegex = new RegExp(/(\d*d\d+(\s*[+-]?\s*\d*d*\d*)?)/, "g");
     const formulaMatch = diceKey.match(diceRegex);
@@ -223,8 +176,7 @@ function buildTable(parsedTable, keys, diceKeys, tableName, parentName) {
       Object.entries(entry).forEach(([key, value]) => {
         if (key === diceKey) {
           result.range = getDiceTableRange(value);
-        }
-        else if (diceKeys.includes(key)) return;
+        } else if (diceKeys.includes(key)) return;
         if (concatKeys) {
           if (result.text != "") result.text += "\n\n";
           result.text += `<b>${key}</b>${value}`;
@@ -238,9 +190,6 @@ function buildTable(parsedTable, keys, diceKeys, tableName, parentName) {
       table.results.push(result);
     });
 
-    console.log(`Generated table entry ${realName}`);
-    // const realTable = new RollTable(table);
-    // generatedTables.push(realTable);
     generatedTables.push(table);
 
   });
@@ -248,61 +197,59 @@ function buildTable(parsedTable, keys, diceKeys, tableName, parentName) {
   return generatedTables;
 }
 
-
 export function generateTable(parentName, html, updateExisting) {
   const document = utils.htmlToDoc(html);
   const tableNodes = document.querySelectorAll("table");
   let tablesMatched = [];
+  let updatedDocument = utils.htmlToDoc(html);
 
   let tableNum = 0;
   tableNodes.forEach((node) => {
+  // for (let i = 0; i < tableNodes.length; i++) {
+    // const node = tableNodes[i];
     const parsedTable = parseTable(node);
     const keys = getHeadings(node);
     const diceKeys = findDiceColumns(node);
-    // const contentChunkId = node.getAttribute("data-content-chunk-id");
     let nameGuess = guessTableName(parentName, document, tableNum);
-
-    console.warn(`Table detection triggered for ${parentName}!`);
-    console.log(`Table: "${nameGuess}"`);
-    console.log(`Dice Keys: ${diceKeys.join(", ")}`);
-    console.log(`Keys: ${keys.join(", ")}`);
-
-    // if (config.debug) console.log(node.outerHTML);
-    // if (config.debug && parsedTable) console.log(parsedTable);
-    // if (parsedTable) console.log(parsedTable);
-
     if (nameGuess.split(" ").length > 5 && diceKeys.length === 1 && keys.length === 2) {
       nameGuess = keys[1];
     }
-
     const finalName = `${parentName}: ${nameGuess}`;
 
-    const builtTables = buildTable(parsedTable, keys, diceKeys, finalName, parentName);
-    if (builtTables.length > 0) {
+    logger.debug(`Table detection triggered for ${parentName}!`);
+    logger.debug(`Table: "${finalName}"`);
+    logger.debug(`Dice Keys: ${diceKeys.join(", ")}`);
+    logger.debug(`Keys: ${keys.join(", ")}`);
 
-      updateCompendium("tables", { tables: builtTables }, updateExisting).then((compendiumTables) => {
-        tablesMatched.push({
-          nameGuess,
-          finalName,
-          parentName,
-          tableNum,
-          length: parsedTable.length,
-          keys: keys,
-          diceKeys: diceKeys,
-          diceTable: diceKeys.length > 0,
-          multiDiceKeys: diceKeys.length > 1,
-          diceKeysNumber: diceKeys.length,
-          totalKeys: keys.length,
-          builtTables,
-          compendiumTables,
-        });
-      });
+    const builtTables = buildTable(parsedTable, keys, diceKeys, finalName, parentName);
+
+    if (builtTables.length > 0) {
+      // these updates are done async, and we continue. this is fine as we actually use the table name for linking
+      const compendiumTables = updateCompendium("tables", { tables: builtTables }, updateExisting);
+
+      let tableData = {
+        nameGuess,
+        finalName,
+        parentName,
+        tableNum,
+        length: parsedTable.length,
+        keys: keys,
+        diceKeys: diceKeys,
+        diceTable: diceKeys.length > 0,
+        multiDiceKeys: diceKeys.length > 1,
+        diceKeysNumber: diceKeys.length,
+        totalKeys: keys.length,
+        builtTables,
+        compendiumTables,
+      };
+      tablesMatched.push(tableData);
+      updatedDocument = tableReplacer(updatedDocument, tableNum, tableData.builtTables);
+
     }
     tableNum++;
+  // }
   });
 
-  if (tablesMatched.length > 0) console.warn(tablesMatched);
-
-  return tablesMatched;
+  return updatedDocument.body.innerHTML;
 
 }